move problem-with-service controller into root directory
